# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TrossApp API",
    "version": "1.0.0",
    "description": "\n        TrossApp Backend REST API - Skills-based work order management system.\n        \n        This documentation covers foundational authentication, user management, \n        and role-based access control endpoints.\n        \n        ## Authentication\n        \n        Most endpoints require a Bearer token in the Authorization header:\n        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n        \n        ## Dual Authentication Strategy\n        \n        - **Development Mode:** JWT-based authentication with test users\n        - **Production Mode:** Auth0 OAuth2/OIDC integration\n        \n        ## Response Format\n        \n        All successful responses follow this format:\n        ```json\n        {\n          \"success\": true,\n          \"data\": {...},\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n        \n        Error responses:\n        ```json\n        {\n          \"success\": false,\n          \"error\": \"Error message describing what went wrong\",\n          \"timestamp\": \"2025-11-12T12:00:00.000Z\"\n        }\n        ```\n      ",
    "contact": {
      "name": "TrossApp Team",
      "email": "dev@trossapp.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
...
    },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/dev/token (dev) or Auth0 (prod)"
      }
    },
...
  },
...
  "paths": {
...
    "/api/inventory": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Get all inventory items",
        "description": "Technicians and above can view all inventory. No row-level security.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory retrieved successfully"
          },
          "403": {
            "description": "Forbidden - Technician+ access required"
          }
        }
      },
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Create new inventory item (dispatcher+ only)",
        "description": "Add a new item to inventory. Dispatcher+ access required.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "sku"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 255
                  },
                  "sku": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inventory item created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid data or duplicate SKU"
          },
          "403": {
            "description": "Forbidden - Dispatcher+ access required"
          }
        }
      }
    },
    "/api/inventory/{id}": {
      "patch": {
        "tags": [
          "Inventory"
        ],
        "summary": "Update inventory item (dispatcher+ only)",
        "description": "Update inventory item details. Dispatcher+ access required.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Inventory item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "sku": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string"
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventory item updated successfully"
          },
          "400": {
            "description": "Bad Request - At least one field required"
          },
          "403": {
            "description": "Forbidden - Dispatcher+ access required"
          },
          "404": {
            "description": "Inventory item not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Inventory"
        ],
        "summary": "Delete inventory item (manager+ only)",
        "description": "Soft delete an inventory item (sets is_active=false). Manager+ access required.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Inventory item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory item deleted successfully"
          },
          "403": {
            "description": "Forbidden - Manager+ access required"
          },
          "404": {
            "description": "Inventory item not found"
          }
        }
      }
    },
...
  }
}